Sonar new-code round: mechanical fixes (bug + blockers + smells)
- S2119 (BUG) Pseudonymizer: one shared static SecureRandom, re-used for the
  blank-salt per-instance key (was a new SecureRandom per construction).
- S1845 (BLOCKER x2) DcsWriter.TitlePlan: factories renamed value->of,
  localized->ofLocalized so they no longer clash with the same-named fields
  (same holder-factory-rename class as the earlier error->failed fixes).
- S1192 (x10): error-message stems extracted to constants (DcsWriter ERR_DATA_SET/
  ERR_PARAMETER/ERR_FIELD_ROLE/ERR_NEEDS_NAME, SpreadsheetTemplateWriter ERR_CELL,
  PiiDefaultsFetcher ERR_REJECTED_RULE, ModifyMetadataTool ERR_IS_A/ERR_NO_BM_MANAGER/
  ERR_NO_BM_MODEL, + reuse KEY_CONTENT), with NON-NLS markers renumbered.
- S1905 (x9): redundant long casts removed (EventLogReader offset/limit/windowCap,
  LgpParser date-key arithmetic - the literals are already long).
- S4276: EventLogQuery.normalise takes UnaryOperator<String> instead of
  Function<String,String>.
- S125 (x12): reworded comments the commented-out-code heuristic tripped on -
  LgfParser per-case format docs (brace token-lists -> prose field lists),
  McpProtocolHandler PII note, ModifyMetadataTool top-climb notes, PiiRedactor
  truncated/fullLength note. No meaning lost; none were dead code.
- S135 (x2): LgpParser.stream reduced to a single loop exit (filter fold);
  MetadataPropertyIntrospector extInfo loop second continue inverted.
- S107: ModifyMetadataTool.dispatchTemplatePayload 8->7 params (hasTemplatePayload
  is derivable from templateSpec != null); CellPlan ctor NOSONAR (it IS the
  parameter object, mirrors the JSON cell spec 1:1).
- S1450: PrivacyTab.disposed NOSONAR (cross-method lifecycle guard, set in
  dispose(), read by the async repo-fetch callback).

Behaviour unchanged; build green.

+++ mcp/bundles/com.ditrix.edt.mcp.server/src/com/ditrix/edt/mcp/server/utils/eventlog/EventLogReader.java
@@ -156,10 +156,10 @@ public final class EventLogReader
         LgpParser lgp = new LgpParser(refs);
         Predicate<EventRecord> filter = q.asPredicate();
 
-        long offset = Math.max(0L, (long)q.offset);
-        long limit = Math.max(0L, (long)q.limit);
+        long offset = Math.max(0L, q.offset);
+        long limit = Math.max(0L, q.limit);
         long windowEnd = offset + limit;
-        int windowCap = (int)Math.min((long)Integer.MAX_VALUE, Math.max(1L, windowEnd));
+        int windowCap = (int)Math.min(Integer.MAX_VALUE, Math.max(1L, windowEnd));
 
         // Window buffering, both bounded to O(offset+limit) (never the whole log):
         //  - ascending: partitions oldest-first; keep the matches at ordinals
